Add .gitignore and LICENSE files, update README.md and environment.yml
This commit adds the .gitignore file to ignore certain files and directories in the repository. It also includes the LICENSE file with the Creative Commons Attribution-NonCommercial-ShareAlike 4.0 International Public License. The README.md file has been updated with a new title and introduction, as well as additional sections for requirements, installation, usage, and more information. Lastly, the environment.yml file has been created to specify the dependencies for the project.